requirements.txt
torch
transformers
Pillow
PyMuPDF
chromadb
groq
streamlit
python-dotenv